Reworked listeners to avoid race-conditions
Now we have a separate FIFO list for result listeners.
Also tweaked timing and avoid to talk over the receiver.
Final suspicion is that we can get powersave indicator if we accidentally
talk over the receiver results. If this is true, then we can solve
it by reissuing the last request.